Rename repository_archive -> collection_artifact_archive
Since artifacts are now the only support archive types now
and they are also no longer associated with galaxy Repository
